placeholders for where errors can be thrown
One possible design for error handling is to throw whenever an error is
taken by default. A differently named take function could take errors
without throwing to pass them along. Because not all exceptions are
error types, complimentary functions could include addErrorEvent and
addNodeCallback functions that will wrap the error message if it is not
of type Error.